+++ src/networks/mainnet/mod.rs
@@ -4,7 +4,10 @@
 use crate::{
     eth::EthChainId,
     make_height,
-    shim::{clock::ChainEpoch, version::NetworkVersion},
+    shim::{
+        clock::{ChainEpoch, EPOCHS_IN_DAY},
+        version::NetworkVersion,
+    },
 };
 use ahash::HashMap;
 use cid::Cid;
@@ -75,9 +78,11 @@ pub static HEIGHT_INFOS: Lazy<HashMap<Height, HeightInfo>> = Lazy::new(|| {
         make_height!(Waffle, 4_154_640, get_bundle_cid("v14.0.0")),
         // Wed 20 Nov 23:00:00 UTC 2024
         make_height!(TukTuk, 4_461_240, get_bundle_cid("v15.0.0")),
-        // TODO(forest): https://github.com/ChainSafe/forest/issues/5041
-        make_height!(Teep, i64::MAX, get_bundle_cid("v16.0.0-rc3")),
-        make_height!(Tock, i64::MAX),
+        // Thu 10 Apr 23:00:00 UTC 2025
+        make_height!(Teep, 4_867_320, get_bundle_cid("v16.0.0")),
+        // This epoch, 90 days after Teep is the completion of FIP-0100 where actors will start applying
+        // the new daily fee to pre-Teep sectors being extended.
+        make_height!(Tock, 4_867_320 + EPOCHS_IN_DAY * 90),
     ])
 });